Add PDF support (#1006)
* Add PDF support

For rendering PDF, we use PDF.js to render PDF in a WebView. To provide
PDF doc (which could be quite big) we stream it via the
Android<->JavaScript bridge, PDF.js requests chunks it needs, and we
serve them from the Android side through that bridge.

So, currently it works this way:

We check in WebFragment's `downloadListener` if the mime type is
`application/pdf`, and if yes - we download the PDF file, and then start
rendering the PDF loader (just an HTML file, that initializes PDF.js and
hooks up the bridge, and also exposes some additional API (e.g. for
zoom).

The bridge provides access to the downloaded file, via
`RandomAccessFile`. We need to support random access.

PDF.js render itself into a container you provide. Built-in PDF.js
`PdfViewer` uses the container's height to figure out what pages it
needs to load/render.  Since in Materialistic the `WebView`'s height
wraps the content, we should use window's height instead of container's
height for that. So, we inherit our custom `CustomPdfViewer` from
`PdfViewer` and override that functionality.

The only HTML/JS stuff I wrote is in `assets/pdf/index.html` and
`assets/pdf/script.js`. All other stuff in `assets/pdf/vendor/pdf.js/` is
taken from the PDF.js repo without any changes.

* Split PDF vendor assets to debug/release non-minified/minified versions

* Create FileDownloader and provide it from NetworkModule